pyproject.toml
[project]
name = "wedding-photos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
dependencies = [
    "boto3>=1.34.0",
    "open-clip-torch>=2.24.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "streamlit>=1.35.0",
    "supabase>=2.4.0",
    "torch>=2.2.0",
]
